@@ -4,13 +4,16 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   cupertino_icons: ^0.1.2
+  geolocator: ^8.1.1
+  http: ^0.13.4
+  flutter_spinkit: ^5.1.0
 
 dev_dependencies:
   flutter_test:
